+++ modules/ui/fields/roadheight.js
@@ -3,7 +3,7 @@ import { select as d3_select } from 'd3-selection';
 import * as countryCoder from '@rapideditor/country-coder';
 
 import { uiCombobox } from '../combobox';
-import { t } from '../../core/localizer';
+import { t, localizer } from '../../core/localizer';
 import { utilGetSetValue, utilNoAuto, utilRebind, utilTotalExtent } from '../../util';
 
 
@@ -16,6 +16,8 @@ export function uiFieldRoadheight(field, context) {
     var _entityIDs = [];
     var _tags;
     var _isImperial;
+    var formatFloat = localizer.floatFormatter(localizer.languageCode());
+    var parseLocaleFloat = localizer.floatParser(localizer.languageCode());
 
     var primaryUnits = [
         {
@@ -129,16 +131,23 @@ export function uiFieldRoadheight(field, context) {
 
         if (!primaryValue && !secondaryValue) {
             tag[field.key] = undefined;
-        } else if (isNaN(primaryValue) || isNaN(secondaryValue) || !_isImperial) {
-            tag[field.key] = context.cleanTagValue(primaryValue);
         } else {
-            if (primaryValue !== '') {
-                primaryValue = context.cleanTagValue(primaryValue + '\'');
+            var rawPrimaryValue = parseLocaleFloat(primaryValue);
+            if (isNaN(rawPrimaryValue)) rawPrimaryValue = primaryValue;
+            var rawSecondaryValue = parseLocaleFloat(secondaryValue);
+            if (isNaN(rawSecondaryValue)) rawSecondaryValue = secondaryValue;
+
+            if (isNaN(rawPrimaryValue) || isNaN(rawSecondaryValue) || !_isImperial) {
+                tag[field.key] = context.cleanTagValue(rawPrimaryValue);
+            } else {
+                if (rawPrimaryValue !== '') {
+                    rawPrimaryValue = rawPrimaryValue + '\'';
+                }
+                if (rawSecondaryValue !== '') {
+                    rawSecondaryValue = rawSecondaryValue + '"';
+                }
+                tag[field.key] = context.cleanTagValue(rawPrimaryValue + rawSecondaryValue);
             }
-            if (secondaryValue !== '') {
-                secondaryValue = context.cleanTagValue(secondaryValue + '"');
-            }
-            tag[field.key] = primaryValue + secondaryValue;
         }
 
         dispatch.call('change', this, tag);
@@ -156,26 +165,36 @@ export function uiFieldRoadheight(field, context) {
             if (primaryValue && (primaryValue.indexOf('\'') >= 0 || primaryValue.indexOf('"') >= 0)) {
                 secondaryValue = primaryValue.match(/(-?[\d.]+)"/);
                 if (secondaryValue !== null) {
-                    secondaryValue = secondaryValue[1];
+                    secondaryValue = formatFloat(parseFloat(secondaryValue[1]));
                 }
                 primaryValue = primaryValue.match(/(-?[\d.]+)'/);
                 if (primaryValue !== null) {
-                    primaryValue = primaryValue[1];
+                    primaryValue = formatFloat(parseFloat(primaryValue[1]));
                 }
                 _isImperial = true;
             } else if (primaryValue) {
+                var rawValue = primaryValue;
+                primaryValue = parseFloat(rawValue);
+                if (isNaN(primaryValue)) {
+                    primaryValue = rawValue;
+                } else {
+                    primaryValue = formatFloat(primaryValue);
+                }
                 _isImperial = false;
             }
         }
 
         setUnitSuggestions();
 
+        // If feet are specified but inches are omitted, assume zero inches.
+        var inchesPlaceholder = formatFloat(0);
+
         utilGetSetValue(primaryInput, typeof primaryValue === 'string' ? primaryValue : '')
             .attr('title', isMixed ? primaryValue.filter(Boolean).join('\n') : null)
             .attr('placeholder', isMixed ? t('inspector.multiple_values') : t('inspector.unknown'))
             .classed('mixed', isMixed);
         utilGetSetValue(secondaryInput, typeof secondaryValue === 'string' ? secondaryValue : '')
-            .attr('placeholder', isMixed ? t('inspector.multiple_values') : (_isImperial ? '0' : null))
+            .attr('placeholder', isMixed ? t('inspector.multiple_values') : (_isImperial ? inchesPlaceholder : null))
             .classed('mixed', isMixed)
             .classed('disabled', !_isImperial)
             .attr('readonly', _isImperial ? null : 'readonly');